Ignore junk output from solvers
In the special case that a formula does not have clauses (because it was
a tautology, perhaps), the solvers return a single line containing "0".
This did not match the regex and an expression involving, consequently,
Nil.starts-with('-') was returned leading to an error. Now only matching
lines are processed.

This drove me crazy while debugging another project using these solvers.
The backtrace referenced a starts-with invocation on Nil, but nothing
of that sort was anywhere near the code indicated in the backtrace
However, that code was, in one way or another, reading the supply given
by this module, which caused the faulty code to execute. (And moreover,
the other project was using the wrong solver -- this one -- which can't
handle empty inputs).

Later this should be a proper error. Error types are a TODO in the main
SAT module.